Remove FAB related auth methods from security manager to auth manager
